+++ apps/api/src/app/import/import-data.dto.ts
@@ -1,7 +1,11 @@
+import { CreateOrderDto } from '@ghostfolio/api/app/order/create-order.dto';
 import { Order } from '@prisma/client';
-import { IsArray } from 'class-validator';
+import { Type } from 'class-transformer';
+import { IsArray, ValidateNested } from 'class-validator';
 
 export class ImportDataDto {
   @IsArray()
-  orders: Partial<Order>[];
+  @Type(() => CreateOrderDto)
+  @ValidateNested({ each: true })
+  orders: Order[];
 }