+++ apps/client/src/app/pages/portfolio/transactions/import-transaction-dialog/interfaces/interfaces.ts
@@ -1,0 +1,4 @@
+export interface ImportTransactionDialogParams {
+  deviceType: string;
+  messages: string[];
+}