+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Extended the validation of the import functionality for transactions
+  - Valid data types
+  - Maximum number of orders
+  - Data provider service returns data for the `dataSource` / `symbol` pair
+
 ### Fixed
 
 - Fixed the broken line charts showing value labels